+++ backend/src/Request/Admin/CaptainFinancialSystem/AdminCaptainFinancialSystemAccordingOnOrderCreateRequest.php
@@ -6,9 +6,9 @@ class AdminCaptainFinancialSystemAccordingOnOrderCreateRequest
 {
     private string $categoryName;
     
-    private int $countKilometersFrom;
+    private float $countKilometersFrom;
 
-    private int $countKilometersTo;
+    private float $countKilometersTo;
 
     private float $amount;
